bug fix 3 commit
