+++ src/features/tableBookings/models/bookingMockData.ts
@@ -1,0 +1,36 @@
+import { type bookingData } from "./types"
+
+export const bookingMockData: bookingData[] = [
+    {
+        id: 8,
+        passengerId: 21,
+        bookingDate: "2022-11-26T21:00:00",
+        bookingStatus: "PAID",
+        flightSeatId: 346,
+        flightId: 349
+    },
+    {
+        id: 9,
+        passengerId: 22,
+        bookingDate: "2022-11-26T21:00:00",
+        bookingStatus: "PAID",
+        flightSeatId: 347,
+        flightId: 350
+    }, 
+    {
+        id: 10,
+        passengerId: 23,
+        bookingDate: "2022-11-26T21:00:00",
+        bookingStatus: "PAID",
+        flightSeatId: 348,
+        flightId: 351   
+    },
+    {
+        id: 11,
+        passengerId: 24,
+        bookingDate: "2022-11-26T21:00:00",
+        bookingStatus: "PAID",
+        flightSeatId: 349,
+        flightId: 352
+    }
+]
