+++ src/features/tableBookings/models/types.ts
@@ -1,0 +1,9 @@
+export interface bookingData {
+    id: number,
+    passengerId: number,
+    bookingDate: string,
+    bookingStatus: string,
+    flightSeatId: number,
+    flightId: number
+  }
+
