+++ src/shared/types/dataTable.ts
@@ -92,11 +92,4 @@ export interface IContentDestinationTable {
   airportName: string;
 }
 
-export interface bookingData {
-  id: number,
-  passengerId: number,
-  bookingDate: string,
-  bookingStatus: string,
-  flightSeatId: number,
-  flightId: number
-}
+
